% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxtracter.R
\name{letter_num}
\alias{letter_num}
\title{Convert xlsx to a dataframe
This function converts a column letter to number}
\usage{
letter_num(ref)
}
\arguments{
\item{ref}{letter for reference}
}
\value{
number
}
\description{
Convert xlsx to a dataframe
This function converts a column letter to number
}
