% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxtracter.R
\name{xlsxtractor}
\alias{xlsxtractor}
\title{This function loads a xlsx file and extracts the passed list to a dataframe}
\usage{
xlsxtractor(infile, sheet, lcolrow)
}
\arguments{
\item{infile}{Path to the input xlsx file}

\item{sheet}{Sheet name}

\item{lcolrow}{List of header name to use, columns, rows, extract type,
cbind or rbind to extract}
}
\value{
A data frame
}
\description{
This function loads a xlsx file and extracts the passed list to a dataframe
}
