% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxtracter.R
\name{xlsxtractor}
\alias{xlsxtractor}
\title{This function loads a xlsx file and extracts the passed list to a dataframe}
\usage{
xlsxtractor(infile, colrow, sheet = "Sheet1")
}
\arguments{
\item{infile}{Path to the input xlsx file}

\item{colrow}{List of header name to use, columns, rows, startrow,
row interval}

\item{sheet}{Optional sheet name}
}
\value{
A data frame
}
\description{
This function loads a xlsx file and extracts the passed list to a dataframe
}
